Rename test files to start with "test"
This allows the default discover on unittest to work
