+++ test/TestArchiveCommand.py
@@ -17,8 +17,8 @@
 import unittest
 
 from topydo.commands.ArchiveCommand import ArchiveCommand
-from test.CommandTest import CommandTest
-from test.TestFacilities import load_file_to_todolist
+from test.TestCommand import CommandTest
+from test.Facilities import load_file_to_todolist
 from topydo.lib.TodoList import TodoList
 
 class ArchiveCommandTest(CommandTest):
