+++ test/TestImportance.py
@@ -14,22 +14,31 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from six import u
+from datetime import date
 import unittest
 
-from test.TestFacilities import load_file
-from test.TopydoTest import TopydoTest
+from topydo.lib.Config import config
+from topydo.lib.Importance import importance
+from topydo.lib.Todo import Todo
+from test.TestTopydo import TopydoTest
 
-class TodoFileTest(TopydoTest):
-    def test_empty_file(self):
-        todofile = load_file('test/data/TodoFileTest1.txt')
+class ImportanceTest(TopydoTest):
+    def test_importance1(self):
+        todo = Todo("Foo")
+        self.assertEqual(importance(todo), 2)
 
-        self.assertEqual(len(todofile), 0)
+    def test_importance2(self):
+        todo = Todo("(A) Foo")
+        self.assertEqual(importance(todo), 5)
 
-    def test_utf_8(self):
-        todofile = load_file('test/data/utf-8.txt')
+    def test_importance3(self):
+        todo = Todo("(A) Foo " + config().tag_star() + ":1")
+        self.assertEqual(importance(todo), 6)
 
-        self.assertEqual(todofile[0].source(), u('(C) \u25ba UTF-8 test \u25c4'))
+    def test_importance4(self):
+        today_str = date.today().isoformat()
+        todo = Todo("(C) Foo " + config().tag_due() + ":" + today_str)
+        self.assertEqual(importance(todo), 8)
 
 if __name__ == '__main__':
     unittest.main()
