+++ test/TestListProjectCommand.py
@@ -17,8 +17,8 @@
 import unittest
 
 from topydo.commands.ListProjectCommand import ListProjectCommand
-from test.CommandTest import CommandTest
-from test.TestFacilities import load_file_to_todolist
+from test.TestCommand import CommandTest
+from test.Facilities import load_file_to_todolist
 
 class ListProjectCommandTest(CommandTest):
     def test_projects1(self):
