+++ test/TestPostponeCommand.py
@@ -19,7 +19,7 @@ import unittest
 from six import u
 
 from topydo.commands.PostponeCommand import PostponeCommand
-from test.CommandTest import CommandTest
+from test.TestCommand import CommandTest
 from topydo.lib.TodoList import TodoList
 
 class PostponeCommandTest(CommandTest):
