+++ test/TestSortCommand.py
@@ -18,8 +18,8 @@ import unittest
 
 from topydo.lib.Config import config
 from topydo.commands.SortCommand import SortCommand
-from test.CommandTest import CommandTest
-from test.TestFacilities import load_file_to_todolist
+from test.TestCommand import CommandTest
+from test.Facilities import load_file_to_todolist
 
 class SortCommandTest(CommandTest):
     def setUp(self):
