+++ test/TestSorter.py
@@ -19,8 +19,8 @@ import unittest
 from topydo.lib.Config import config
 from topydo.lib.Sorter import Sorter
 
-from test.TestFacilities import load_file, todolist_to_string, load_file_to_todolist, print_view
-from test.TopydoTest import TopydoTest
+from test.Facilities import load_file, todolist_to_string, load_file_to_todolist, print_view
+from test.TestTopydo import TopydoTest
 
 class SorterTest(TopydoTest):
     def sort_file(self, p_filename, p_filename_ref, p_sorter):
