+++ test/TestTagCommand.py
@@ -17,7 +17,7 @@
 import unittest
 
 from topydo.commands.TagCommand import TagCommand
-from test.CommandTest import CommandTest
+from test.TestCommand import CommandTest
 from topydo.lib.TodoList import TodoList
 
 class TagCommandTest(CommandTest):
