+++ test/TestTodoFile.py
@@ -14,24 +14,22 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+from six import u
 import unittest
 
-from topydo.lib.Utils import escape_ansi
-from test.TopydoTest import TopydoTest
+from test.Facilities import load_file
+from test.TestTopydo import TopydoTest
 
-class CommandTest(TopydoTest):
-    def __init__(self, *args, **kwargs):
-        super(CommandTest, self).__init__(*args, **kwargs)
-        self.output = ""
-        self.errors = ""
+class TodoFileTest(TopydoTest):
+    def test_empty_file(self):
+        todofile = load_file('test/data/TodoFileTest1.txt')
 
-    def out(self, p_output):
-        if p_output:
-            self.output += escape_ansi(p_output + "\n")
+        self.assertEqual(len(todofile), 0)
 
-    def error(self, p_error):
-        if p_error:
-            self.errors += escape_ansi(p_error + "\n")
+    def test_utf_8(self):
+        todofile = load_file('test/data/utf-8.txt')
+
+        self.assertEqual(todofile[0].source(), u('(C) \u25ba UTF-8 test \u25c4'))
 
 if __name__ == '__main__':
     unittest.main()
