+++ test/TestView.py
@@ -18,10 +18,10 @@ import unittest
 
 from topydo.lib import Filter
 from topydo.lib.Sorter import Sorter
-from test.TestFacilities import load_file, todolist_to_string, print_view
+from test.Facilities import load_file, todolist_to_string, print_view
 from topydo.lib.TodoFile import TodoFile
 from topydo.lib.TodoList import TodoList
-from test.TopydoTest import TopydoTest
+from test.TestTopydo import TopydoTest
 
 class ViewTest(TopydoTest):
     def test_view(self):
